+++ node/composable-example/clean.js
@@ -4,9 +4,9 @@ import { fileURLToPath } from "url";
 function getCurrentFolder() {
     const folder = fileURLToPath(new URL(".", import.meta.url));
     // for windows OS, convert backslash to regular slash and drop drive letter from path
-    const path = folder.includes('\\')? folder.replaceAll('\\', '/') : folder;
-    const colon = path.indexOf(':');
-    return colon === 1? path.substring(colon+1) : path;
+    const filePath = folder.includes('\\')? folder.replaceAll('\\', '/') : folder;
+    const colon = filePath.indexOf(':');
+    return colon === 1? filePath.substring(colon+1) : filePath;
 }
 
 function getFolder(target) {