+++ node/composable-example/test/tasks/simple-test-task.ts
@@ -1,0 +1,14 @@
+import { Composable, EventEnvelope } from "mercury-composable";
+
+export class SimpleTestTask implements Composable {
+    static readonly routeName = 'simple.test.task'
+
+    // @preload(SimpleTestTask.routeName, 10)
+    initialize(): Composable {
+        return this;
+    }
+
+    async handleEvent(evt: EventEnvelope) {
+        return evt.setHeader('type', 'simple-test');
+    }
+}