+++ node/minimalist-kafka-adapter/src/workers/simple-consumer.ts
@@ -1,19 +1,29 @@
 import { EventEnvelope, Logger, Utility } from 'mercury-composable';
-import { Kafka, Consumer, EachMessageHandler } from 'kafkajs';
+import { Kafka, Consumer } from 'kafkajs';
+import EventEmitter from 'events';
 
 const log = Logger.getInstance();
 const util = new Utility();
+type MessageHandler = (payload: object) => Promise<void>;
 
 export class SimpleKafkaConsumer {
     private readonly consumer: Consumer;
     private readonly id: string;
+    private readonly emitter: EventEmitter;
+    private seq = 0;
     private connected = false;
 
-    constructor(client: Kafka, groupId: string) {
+    constructor(client: Kafka | EventEmitter, groupId: string) {        
         this.id = util.getUuid();
-        this.consumer = client.consumer({
-            groupId: groupId
-        });
+        if (client instanceof Kafka) {
+            this.consumer = client.consumer({
+                groupId: groupId
+            });
+        } else if (client instanceof EventEmitter) {
+            this.emitter = client;
+        } else {
+            throw new Error('client must be an instance of Kafka or EventEmitter');
+        }
     }
 
     getId() {
@@ -21,18 +31,46 @@ export class SimpleKafkaConsumer {
     }
 
     async connect() {
-        await this.consumer.connect();
-        this.connected = true;
-        log.info(`Consumer ${this.id} started`);
+        if (!this.connected) {
+            if (this.consumer) {
+                await this.consumer.connect();    
+            }
+            log.info(`Consumer ${this.id} started`);
+            this.connected = true; 
+        }
     }
 
-    async subscribe(topics: Array<string>, handleEvent: EachMessageHandler) {
-        if (this.consumer && this.connected) {
-            await this.consumer.subscribe({ topics: topics });
-            log.info(`Consumer ${this.id} subscribed to ${JSON.stringify(topics)}`);
-            await this.consumer.run({
-                eachMessage: handleEvent
-            });            
+    async subscribe(topic: string, handleEvent: MessageHandler) {
+        if (this.connected) {
+            if (this.consumer) {
+                await this.consumer.subscribe({ topics: [topic] });                
+                await this.consumer.run({
+                    eachMessage: handleEvent
+                });
+            } else {
+                // consume message from Kafka emulator
+                this.emitter.on(topic, async (b: Buffer) => {
+                    this.seq++;
+                    const evt = new EventEnvelope(b);
+                    const bytes = evt.getBody() as Buffer;
+                    const message = {
+                        key: Buffer.from(evt.getId()),
+                        value: bytes,
+                        timestamp: new Date().toISOString(),
+                        attributes: 0,
+                        offset: String(this.seq),
+                        size: b.length,
+                        headers: evt.getHeaders()
+                    }
+                    const data = {
+                        topic: topic,
+                        partition: 0,
+                        message: message
+                    };
+                    await handleEvent(data);
+                });                
+            }
+            log.info(`Consumer ${this.id} subscribed to ${topic}`);           
         } else {
             throw new Error(`Consumer ${this.id} not ready`);
         }
@@ -54,9 +92,11 @@ export class SimpleKafkaConsumer {
     }
 
     async close() {
-        if (this.consumer && this.connected) {
-            await this.consumer.disconnect();
-            log.info(`Consumer ${this.id} stopped`);    
+        if (this.connected) {
+            if (this.consumer) {
+                await this.consumer.disconnect();
+            }
+            log.info(`Consumer ${this.id} stopped`);
         }
     }
 }